+++ changelog.md
@@ -1,4 +1,21 @@
 # Changelog
 
-**./src/common/value-prep.js**
-* 'checkInputObjectType' now correctly checks for 'object' type instead of 'string'.
+**maze.js**
+* Defined three new commands:
+	* 'absolute-to-relative
+	* grid-to-relative
+	* image-to-relative
+* These convert the respective input into a Relative definition text file.
+* Only the stubs have been written for now.
+
+---
+
+**./src/common/sub-interface/cmd-desc.js**
+* Added new function 'writeRelativeConversionDesc'
+	* Writes description for the `%type%-to-relative` commands.
+* Wrote descriptions for new commands:
+	* absoluteToRelativeDesc
+	* gridToRelativeDesc
+	* imageToRelativeDesc
+
+